Refactor sub components
