+++ src/main/java/seedu/programmer/ui/components/CommandHistory.java
@@ -1,4 +1,4 @@
-package seedu.programmer.ui;
+package seedu.programmer.ui.components;
 
 import java.util.ArrayList;
 import java.util.List;
