+++ src/main/java/seedu/programmer/ui/components/ResultDisplay.java
@@ -1,10 +1,11 @@
-package seedu.programmer.ui;
+package seedu.programmer.ui.components;
 
 import static java.util.Objects.requireNonNull;
 
 import javafx.fxml.FXML;
 import javafx.scene.control.TextArea;
 import javafx.scene.layout.Region;
+import seedu.programmer.ui.UiPart;
 
 /**
  * A UI for the status bar that is displayed at the header of the application.
