+++ src/main/java/seedu/programmer/ui/windows/HelpWindow.java
@@ -1,4 +1,4 @@
-package seedu.programmer.ui;
+package seedu.programmer.ui.windows;
 
 import javafx.beans.property.SimpleStringProperty;
 import javafx.collections.FXCollections;
